+++ src/main.rs
@@ -6,5 +6,4 @@ async fn main() -> ElusionResult<()> {
     println!("HELLO, ELUSION!!!");
 
     Ok(())
-}
- 
+}